+++ headers/RGBCell.h
@@ -1,22 +1,24 @@
 #ifndef RGBCELL_H
 #define RGBCELL_H
 
+#include <cstdint>
+
 class RGBCell
 {
 private:
-    int _redColor;
-    int _greenColor;
-    int _blueColor;
+    uint8_t _redColor;
+    uint8_t _greenColor;
+    uint8_t _blueColor;
 public:
-    RGBCell(int r, int g, int b);
+    RGBCell(uint8_t r, uint8_t g, uint8_t b);
     RGBCell();
-    int getRedValue();
-    int getGreenValue();
-    int getBlueValue();
+    uint8_t getRedValue();
+    uint8_t getGreenValue();
+    uint8_t getBlueValue();
 
-    void setRedValue(int value);
-    void setGreenValue(int value);
-    void setBlueValue(int value);
+    void setRedValue(uint8_t value);
+    void setGreenValue(uint8_t value);
+    void setBlueValue(uint8_t value);
 };
 
 #endif //RGBCELL_H
